22_VLM_For_Better_Understanding/requirements.txt
torch
transformers
Pillow
beautifulsoup4
matplotlib
pdf2image
timm
flask
jsons
accelerate
flask-cors
opencv-python
requests
